refactor codes and remove unnecessary configs/codes
