+++ test/statsd/ecs_resources/ec2_launch/daemon/ecs_taskdef.tpl
@@ -7,10 +7,6 @@
       {
         "name": "CW_CONFIG_CONTENT",
         "valueFrom": "${cwagent_ssm_parameter_arn}"
-      },
-      {
-        "name": "PROMETHEUS_CONFIG_CONTENT",
-        "valueFrom": "${prometheus_ssm_parameter_arn}"
       }
     ],
     "logConfiguration": {
@@ -69,7 +65,7 @@
       "entryPoint": [
           "/bin/sh",
           "-c",
-          "while true; do echo 'statsd_counter_1:1.0|c|#key:value' | socat -v -t 0 - UDP:cloudwatch_agent:8125; sleep 1; done"
+          "while true; do echo 'statsd_counter_1:1.0|c|#key:value' | socat -v -t 0 - UDP:cloudwatch_agent:8125; echo 'statsd_gauge_1:1.0|g|#key:value' | socat -v -t 0 - UDP:cloudwatch_agent:8125; sleep 1; done"
       ],
       "logConfiguration": {
         "logDriver": "awslogs",
